DataWarehouse: update products controller

+++ DataWarehouse.API/Controllers/ProductController.cs
@@ -1,6 +1,7 @@
 ﻿using DataWarehouse.API.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using Store.Core.Shared;
 using System.Threading.Tasks;
 
 namespace DataWarehouse.API.Controllers
@@ -19,13 +20,19 @@ namespace DataWarehouse.API.Controllers
             _productsService = productsService;
         }
 
+        /// <summary>
+        /// Get the list of all products from MongoDB
+        /// </summary>
         [HttpGet]
-        public async Task<IActionResult> GetAllProductsChanges()
+        public async Task<IActionResult> GetProducts()
         {
             var products = await _productsService.GetProductsAsync();
             return Ok(products);
         }
 
+        /// <summary>
+        /// Get the product by id from MongoDB
+        /// </summary>
         [HttpGet("{id}")]
         public async Task<IActionResult> GetProduct(string id)
         {
@@ -33,28 +40,68 @@ namespace DataWarehouse.API.Controllers
             return Ok(product);
         }
 
+        /// <summary>
+        /// Create a product in MongoDB
+        /// </summary>
         [HttpPost]
-        public async Task<IActionResult> CreateProduct(Product product)
+        public async Task<IActionResult> CreateProduct(ProductDto productDto)
         {
+            var product = new Product
+            {
+                CategoryId = productDto.CategoryId,
+                Price = productDto.Price,
+                ProductId = productDto.EntityId,
+                Description = productDto.Description,
+                ImagePath = productDto.ImagePath,
+                ProviderId = productDto.ProviderId,
+                Name = productDto.Name,
+                Operation = productDto.Operation,
+                Type = productDto.Type
+            };
             await _productsService.CreateAsync(product);
             return Ok();
         }
 
+        /// <summary>
+        /// Delete the product from MongoDB
+        /// </summary>
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteProduct(string id)
         {
+            var product = await _productsService.GetProductAsync(id);
+            if(product == null)
+            {
+                return NotFound();
+            }
             await _productsService.RemoveAsync(id);
             return Ok();
         }
 
+        /// <summary>
+        /// Update the product in MongoDB
+        /// </summary>
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateProduct(Product product)
+        public async Task<IActionResult> UpdateProduct(string id, ProductDto productDto)
         {
+            var product = await _productsService.GetProductAsync(id);
+            if (product == null)
+            {
+                return NotFound();
+            }
+
+            product.CategoryId = productDto.CategoryId;
+            product.Price = productDto.Price;
+            product.ProductId = productDto.EntityId;
+            product.Description = productDto.Description;
+            product.ImagePath = productDto.ImagePath;
+            product.ProviderId = productDto.ProviderId;
+            product.Name = productDto.Name;
+            product.Operation = productDto.Operation;
+            product.Type = productDto.Type;
+            
             await _productsService.UpdateAsync(product);
             return Ok();
         }
-
-
 
     }
 }
